+++ types/three/examples/jsm/nodes/Nodes.d.ts
@@ -151,7 +151,7 @@ export {
 export { default as CondNode, cond } from './math/CondNode.js';
 
 // math utils
-export { parabola, gain, pcurve, sinc } from './math/MathUtilsNode.js';
+export { parabola, gain, pcurve, sinc } from './math/MathUtils.js';
 export { triNoise3D } from './math/TriNoise3D.js';
 
 // utils
